fix #237: static build / link
